# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/bills": {
            "get": {
                "description": "list bill entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List bill entities",
                "operationId": "list-bill",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Bill"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create bill",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create bill",
                "operationId": "create-bill",
                "parameters": [
                    {
                        "description": "Bill entity",
                        "name": "bill",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Bill"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Bill"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/bills/{id}": {
            "get": {
                "description": "get bill by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a bill entity by ID",
                "operationId": "get-bill",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Bill"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get bill by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a bill entity by ID",
                "operationId": "delete-bill",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.Bill": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "financier": {
                    "type": "integer"
                },
                "paytype": {
                    "type": "integer"
                },
                "unpaybill": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Bill": {
            "type": "object",
            "properties": {
                "Amount": {
                    "description": "Amount holds the value of the \"Amount\" field.",
                    "type": "string"
                },
                "Date": {
                    "description": "Date holds the value of the \"Date\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BillQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BillEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.BillEdges": {
            "type": "object",
            "properties": {
                "EdgesOfOfficer": {
                    "description": "EdgesOfOfficer holds the value of the EdgesOfOfficer edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Financier"
                },
                "EdgesOfPaytype": {
                    "description": "EdgesOfPaytype holds the value of the EdgesOfPaytype edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Paytype"
                },
                "EdgesOfTreatment": {
                    "description": "EdgesOfTreatment holds the value of the EdgesOfTreatment edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Unpaybill"
                }
            }
        },
...
        "ent.Doctor": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DoctorQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DoctorEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DoctorEdges": {
            "type": "object",
            "properties": {
                "EdgesOfDoctorinfo": {
                    "description": "EdgesOfDoctorinfo holds the value of the EdgesOfDoctorinfo edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctorinfo"
                },
                "EdgesOfTreatment": {
                    "description": "EdgesOfTreatment holds the value of the EdgesOfTreatment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Treatment"
                    }
                },
                "EdgesOfUser": {
                    "description": "EdgesOfUser holds the value of the EdgesOfUser edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
...
        "ent.Financier": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the FinancierQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.FinancierEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.FinancierEdges": {
            "type": "object",
            "properties": {
                "EdgesOfBills": {
                    "description": "EdgesOfBills holds the value of the EdgesOfBills edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bill"
                    }
                },
                "EdgesOfUser": {
                    "description": "EdgesOfUser holds the value of the EdgesOfUser edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
...
        "ent.Medicalrecordstaff": {
            "type": "object",
            "properties": {
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MedicalrecordstaffQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicalrecordstaffEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.MedicalrecordstaffEdges": {
            "type": "object",
            "properties": {
                "EdgesOfMedicalrecordstaffPatientrights": {
                    "description": "EdgesOfMedicalrecordstaffPatientrights holds the value of the EdgesOfMedicalrecordstaffPatientrights edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Patientrights"
                    }
                },
                "EdgesOfPatientrecord": {
                    "description": "EdgesOfPatientrecord holds the value of the EdgesOfPatientrecord edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Patientrecord"
                    }
                },
                "EdgesOfUser": {
                    "description": "EdgesOfUser holds the value of the EdgesOfUser edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.Nurse": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the NurseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.NurseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "nursinglicense": {
                    "description": "Nursinglicense holds the value of the \"nursinglicense\" field.",
                    "type": "string"
                },
                "position": {
                    "description": "Position holds the value of the \"position\" field.",
                    "type": "string"
                }
            }
        },
        "ent.NurseEdges": {
            "type": "object",
            "properties": {
                "EdgesOfHistorytaking": {
                    "description": "EdgesOfHistorytaking holds the value of the EdgesOfHistorytaking edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Historytaking"
                    }
                },
                "EdgesOfUser": {
                    "description": "EdgesOfUser holds the value of the EdgesOfUser edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
...
        "ent.Patientrecord": {
            "type": "object",
            "properties": {
                "Age": {
                    "description": "Age holds the value of the \"Age\" field.",
                    "type": "integer"
                },
                "Allergic": {
                    "description": "Allergic holds the value of the \"Allergic\" field.",
                    "type": "string"
                },
                "Bloodtype": {
                    "description": "Bloodtype holds the value of the \"Bloodtype\" field.",
                    "type": "string"
                },
                "Date": {
                    "description": "Date holds the value of the \"Date\" field.",
                    "type": "string"
                },
                "Disease": {
                    "description": "Disease holds the value of the \"Disease\" field.",
                    "type": "string"
                },
                "Email": {
                    "description": "Email holds the value of the \"Email\" field.",
                    "type": "string"
                },
                "Home": {
                    "description": "Home holds the value of the \"Home\" field.",
                    "type": "string"
                },
                "Idcardnumber": {
                    "description": "Idcardnumber holds the value of the \"Idcardnumber\" field.",
                    "type": "integer"
                },
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "Phonenumber": {
                    "description": "Phonenumber holds the value of the \"Phonenumber\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientrecordQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientrecordEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Patientrights": {
            "type": "object",
            "properties": {
                "PermissionDate": {
                    "description": "PermissionDate holds the value of the \"PermissionDate\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientrightsQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientrightsEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "insurance_id": {
                    "type": "integer"
                },
                "patientrightstype_id": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Paytype": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PaytypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PaytypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "paytype": {
                    "description": "Paytype holds the value of the \"paytype\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PaytypeEdges": {
            "type": "object",
            "properties": {
                "EdgesOfBills": {
                    "description": "EdgesOfBills holds the value of the EdgesOfBills edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bill"
                    }
                }
            }
        },
...
        "ent.Registrar": {
            "type": "object",
            "properties": {
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RegistrarQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RegistrarEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.RegistrarEdges": {
            "type": "object",
            "properties": {
                "EdgesOfUser": {
                    "description": "EdgesOfUser holds the value of the EdgesOfUser edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
...
        "ent.Treatment": {
            "type": "object",
            "properties": {
                "Datetreat": {
                    "description": "Datetreat holds the value of the \"Datetreat\" field.",
                    "type": "string"
                },
                "Treatment": {
                    "description": "Treatment holds the value of the \"Treatment\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TreatmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TreatmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.TreatmentEdges": {
            "type": "object",
            "properties": {
                "EdgesOfDoctor": {
                    "description": "EdgesOfDoctor holds the value of the EdgesOfDoctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "EdgesOfPatientrecord": {
                    "description": "EdgesOfPatientrecord holds the value of the EdgesOfPatientrecord edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patientrecord"
                },
                "EdgesOfTypetreatment": {
                    "description": "EdgesOfTypetreatment holds the value of the EdgesOfTypetreatment edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Typetreatment"
                },
                "EdgesOfUnpaybills": {
                    "description": "EdgesOfUnpaybills holds the value of the EdgesOfUnpaybills edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Unpaybill"
                }
            }
        },
        "ent.Typetreatment": {
            "type": "object",
            "properties": {
                "Typetreatment": {
                    "description": "Typetreatment holds the value of the \"Typetreatment\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TypetreatmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TypetreatmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Unpaybill": {
            "type": "object",
            "properties": {
                "Status": {
                    "description": "Status holds the value of the \"Status\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UnpaybillQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UnpaybillEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.UnpaybillEdges": {
            "type": "object",
            "properties": {
                "EdgesOfBills": {
                    "description": "EdgesOfBills holds the value of the EdgesOfBills edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Bill"
                },
                "EdgesOfTreatment": {
                    "description": "EdgesOfTreatment holds the value of the EdgesOfTreatment edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Treatment"
                }
            }
        },
        "ent.User": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "images": {
                    "description": "Images holds the value of the \"images\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "EdgesOfDoctor": {
                    "description": "EdgesOfDoctor holds the value of the EdgesOfDoctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "EdgesOfFinancier": {
                    "description": "EdgesOfFinancier holds the value of the EdgesOfFinancier edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Financier"
                },
                "EdgesOfMedicalrecordstaff": {
                    "description": "EdgesOfMedicalrecordstaff holds the value of the EdgesOfMedicalrecordstaff edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Medicalrecordstaff"
                },
                "EdgesOfNurse": {
                    "description": "EdgesOfNurse holds the value of the EdgesOfNurse edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Nurse"
                },
                "EdgesOfUser2registrar": {
                    "description": "EdgesOfUser2registrar holds the value of the EdgesOfUser2registrar edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Registrar"
                },
                "EdgesOfUserPatientrights": {
                    "description": "EdgesOfUserPatientrights holds the value of the EdgesOfUserPatientrights edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patientrights"
                },
                "EdgesOfUserstatus": {
                    "description": "EdgesOfUserstatus holds the value of the EdgesOfUserstatus edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Userstatus"
                }
            }
        },
        "ent.Userstatus": {
            "type": "object",
            "properties": {
                "Userstatus": {
                    "description": "Userstatus holds the value of the \"Userstatus\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserstatusQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserstatusEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.UserstatusEdges": {
            "type": "object",
            "properties": {
                "EdgesOfUser": {
                    "description": "EdgesOfUser holds the value of the EdgesOfUser edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.User"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}